+++ data/examples/other/disabling/empty-disabled-region-out.hs
@@ -1,0 +1,9 @@
+{- ORMOLU_DISABLE -}
+{- ORMOLU_ENABLE -}
+foo = 0
+
+{- ORMOLU_DISABLE -}
+{- ORMOLU_ENABLE -}
+bar = 1
+
+{- ORMOLU_DISABLE -}