+++ data/examples/other/disabling/line-pragma-in-indented-disabled-region.hs
@@ -1,0 +1,8 @@
+foo =
+    {- ORMOLU_DISABLE -}
+    testCase "Foo" testFoo :
+{-# LINE 437 "SomeFile.vhs" #-}
+    testCase "Bar" testBar :
+    testCase "Baz" testBaz :
+    {- ORMOLU_ENABLE -}
+    []