+++ data/examples/other/disabling/normalize-magic-comments-out.hs
@@ -1,0 +1,9 @@
+{- ORMOLU_DISABLE -}
+  {-  ORMOLU_DISABLE   -}
+foo=bar
+{- ORMOLU_ENABLE -}
+{- ORMOLU_ENABLE -}
+bar = baz
+
+{- ORMOLU_DISABLE -}
+baz=foo