+++ src/Ormolu/Parser/Result.hs
@@ -16,6 +16,7 @@ import Ormolu.Parser.Anns
 import Ormolu.Parser.CommentStream
 import Ormolu.Parser.Pragma (Pragma)
 import Ormolu.Parser.Shebang (Shebang)
+import Ormolu.Processing.Disabling (DisabledRegions)
 
 -- | A collection of data that represents a parsed module in Ormolu.
 data ParseResult = ParseResult
@@ -40,7 +41,9 @@ data ParseResult = ParseResult
     -- | Literal suffix
     prLiteralSuffix :: Text,
     -- | Indentation level, can be non-zero in case of region formatting
-    prIndent :: Int
+    prIndent :: Int,
+    -- | Regions where formatting is disabled, cut out from the input
+    prDisabledRegions :: DisabledRegions
   }
 
 -- | Pretty-print a 'ParseResult'.
@@ -49,6 +52,8 @@ prettyPrintParseResult ParseResult {..} =
   unlines
     [ "parse result:",
       "  comment stream:",
-      showCommentStream prCommentStream
+      showCommentStream prCommentStream,
+      "  disabled regions:",
+      show prDisabledRegions
       -- XXX extend as needed
     ]