+++ src/Ormolu/Processing/Preprocess.hs
@@ -11,13 +11,14 @@ where
 import Control.Monad
 import Data.Char (isSpace)
 import qualified Data.List as L
-import Data.Maybe (isJust, maybeToList)
+import Data.Maybe (catMaybes, isJust, maybeToList)
 import GHC.Data.FastString
 import GHC.Types.SrcLoc
 import Ormolu.Config (RegionDeltas (..))
 import Ormolu.Parser.Shebang (isShebang)
-import Ormolu.Processing.Common
 import qualified Ormolu.Processing.Cpp as Cpp
+import Ormolu.Processing.Disabling
+import Ormolu.Utils (getIndentationLevel)
 
 -- | Transform given input possibly returning comments extracted from it.
 -- This handles LINE pragmas, CPP, shebangs, and the magic comments for
@@ -29,25 +30,43 @@ preprocess ::
   String ->
   -- | Region deltas
   RegionDeltas ->
-  -- | Literal prefix, pre-processed input, literal suffix, extra comments
-  (String, String, String, [RealLocated String])
+  -- | Literal prefix, pre-processed input, literal suffix, extra comments,
+  -- disabled regions
+  (String, String, String, [RealLocated String], DisabledRegions)
 preprocess path input RegionDeltas {..} =
-  go 1 OrmoluEnabled Cpp.Outside id id regionLines
+  go 1 OrmoluEnabled Cpp.Outside id id id regionLines
   where
     (prefixLines, otherLines) = splitAt regionPrefixLength (lines input)
     (regionLines, suffixLines) =
       let regionLength = length otherLines - regionSuffixLength
        in splitAt regionLength otherLines
-    go !n ormoluState cppState inputSoFar csSoFar = \case
-      [] ->
-        let input' = unlines (inputSoFar [])
-         in ( unlines prefixLines,
-              case ormoluState of
-                OrmoluEnabled -> input'
-                OrmoluDisabled -> input' ++ endDisabling,
-              unlines suffixLines,
-              csSoFar []
+    finaliseDisabledRegion r =
+      let ls = r []
+          n =
+            getIndentationLevel 0
+              . catMaybes
+              $ fmap
+                ( \case
+                    NoReindent _ -> Nothing
+                    Reindent l -> Just l
+                )
+                ls
+       in fmap
+            ( \case
+                NoReindent l -> NoReindent l
+                Reindent l -> Reindent (drop n l)
             )
+            ls
+    go !n ormoluState cppState inputSoFar csSoFar drsSoFar = \case
+      [] ->
+        ( unlines prefixLines,
+          unlines (inputSoFar []),
+          unlines suffixLines,
+          csSoFar [],
+          drsSoFar $ case ormoluState of
+            OrmoluDisabled r -> [finaliseDisabledRegion r]
+            OrmoluEnabled -> []
+        )
       (x : xs) ->
         let (x', ormoluState', cppState', cs) =
               processLine path n ormoluState cppState x
@@ -57,6 +76,12 @@ preprocess path input RegionDeltas {..} =
               cppState'
               (inputSoFar . (x' :))
               (csSoFar . (maybeToList cs ++))
+              ( case (ormoluState, ormoluState') of
+                  (OrmoluDisabled r, OrmoluEnabled) ->
+                    drsSoFar . (finaliseDisabledRegion r :)
+                  _ ->
+                    drsSoFar
+              )
               xs
 
 -- | Transform a given line possibly returning a comment extracted from it.
@@ -73,35 +98,40 @@ processLine ::
   String ->
   -- | Adjusted line and possibly a comment extracted from it
   (String, OrmoluState, Cpp.State, Maybe (RealLocated String))
-processLine path n ormoluState Cpp.Outside line
-  | "{-# LINE" `L.isPrefixOf` line =
-    let (pragma, res) = getPragma line
-        size = length pragma
-        ss = mkRealSrcSpan (mkRealSrcLoc' 1) (mkRealSrcLoc' (size + 1))
-     in (res, ormoluState, Cpp.Outside, Just (L ss pragma))
-  | isOrmoluEnable line =
-    case ormoluState of
-      OrmoluEnabled ->
-        (enableMarker, OrmoluEnabled, Cpp.Outside, Nothing)
-      OrmoluDisabled ->
-        (endDisabling ++ enableMarker, OrmoluEnabled, Cpp.Outside, Nothing)
-  | isOrmoluDisable line =
-    case ormoluState of
-      OrmoluEnabled ->
-        (disableMarker ++ startDisabling, OrmoluDisabled, Cpp.Outside, Nothing)
-      OrmoluDisabled ->
-        (disableMarker, OrmoluDisabled, Cpp.Outside, Nothing)
-  | isShebang line =
-    let ss = mkRealSrcSpan (mkRealSrcLoc' 1) (mkRealSrcLoc' (length line))
-     in ("", ormoluState, Cpp.Outside, Just (L ss line))
-  | otherwise =
-    let (line', cppState') = Cpp.processLine line Cpp.Outside
-     in (line', ormoluState, cppState', Nothing)
+processLine path n = processLine'
   where
     mkRealSrcLoc' = mkRealSrcLoc (mkFastString path) n
-processLine _ _ ormoluState cppState line =
-  let (line', cppState') = Cpp.processLine line cppState
-   in (line', ormoluState, cppState', Nothing)
+    processCpp ormoluState line cppState =
+      let (line', cppState') = Cpp.processLine line cppState
+       in case ormoluState of
+            OrmoluEnabled ->
+              (line', OrmoluEnabled, cppState', Nothing)
+            OrmoluDisabled r ->
+              ("", OrmoluDisabled $ r . (Reindent line' :), cppState', Nothing)
+    processLine' OrmoluEnabled Cpp.Outside line
+      | isLinePragma line =
+        let (pragma, res) = getPragma line
+            size = length pragma
+            ss = mkRealSrcSpan (mkRealSrcLoc' 1) (mkRealSrcLoc' (size + 1))
+         in (res, OrmoluEnabled, Cpp.Outside, Just (L ss pragma))
+      | isOrmoluEnable line =
+        (enableMarker, OrmoluEnabled, Cpp.Outside, Nothing)
+      | isOrmoluDisable line =
+        (disableMarker, OrmoluDisabled id, Cpp.Outside, Nothing)
+      | isShebang line =
+        let ss = mkRealSrcSpan (mkRealSrcLoc' 1) (mkRealSrcLoc' (length line))
+         in ("", OrmoluEnabled, Cpp.Outside, Just (L ss line))
+      | otherwise =
+        processCpp OrmoluEnabled line Cpp.Outside
+    processLine' (OrmoluDisabled r) Cpp.Outside line
+      | isLinePragma line =
+        ("", OrmoluDisabled $ r . (NoReindent line :), Cpp.Outside, Nothing)
+      | isOrmoluEnable line =
+        (enableMarker, OrmoluEnabled, Cpp.Outside, Nothing)
+      | otherwise =
+        processCpp (OrmoluDisabled r) line Cpp.Outside
+    processLine' ormoluState cppState line =
+      processCpp ormoluState line cppState
 
 -- | Take a line pragma and output its replacement (where line pragma is
 -- replaced with spaces) and the contents of the pragma itself.
@@ -110,20 +140,17 @@ getPragma ::
   String ->
   -- | Contents of the pragma and its replacement line
   (String, String)
-getPragma [] = error "Ormolu.Preprocess.getPragma: input must not be empty"
+getPragma [] =
+  error "Ormolu.Processing.Preprocess.getPragma: input must not be empty"
 getPragma s@(x : xs)
   | "#-}" `L.isPrefixOf` s = ("#-}", "   " ++ drop 3 s)
   | otherwise =
     let (prag, remline) = getPragma xs
      in (x : prag, ' ' : remline)
 
--- | Canonical enable marker.
-enableMarker :: String
-enableMarker = "{- ORMOLU_ENABLE -}"
-
--- | Canonical disable marker.
-disableMarker :: String
-disableMarker = "{- ORMOLU_DISABLE -}"
+-- | Return 'True' if the given string looks like a LINE pragma.
+isLinePragma :: String -> Bool
+isLinePragma = L.isPrefixOf "{-# LINE"
 
 -- | Return 'True' if the given string is an enabling marker.
 isOrmoluEnable :: String -> Bool